mindmap/ai/requirements.txt
fastapi
uvicorn
openai
scikit-learn
numpy
python-dotenv